fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.4
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
idna==3.10
motor==3.6.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
redis==5.2.0
setuptools==75.3.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
websockets==13.1
